feat: login with flatlaf
