+++ StudentManagement.iml
@@ -9,7 +9,6 @@
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" exported="" name="json-20240303" level="project" />
-
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="module-library">
@@ -62,6 +61,15 @@
       <library>
         <CLASSES>
           <root url="jar://$MODULE_DIR$/lib/okio-2.10.0.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library" exported="">
+      <library>
+        <CLASSES>
+          <root url="jar://$USER_HOME$/Downloads/miglayout-swing-11.4.jar!/" />
         </CLASSES>
         <JAVADOC />
         <SOURCES />
