--- src/com/TLU/studentmanagement/Main.java
@@ -1,11 +1,0 @@
-package com.TLU.studentmanagement;
-
-import com.TLU.studentmanagement.controller.LoginController;
-import com.TLU.studentmanagement.controller.RegisterController;
-import com.TLU.studentmanagement.util.ApiClient;
-
-public class Main {
-    public static void main(String[] args) {
-
-    }
-}
