--- src/com/TLU/studentmanagement/controller/LoginController.java
@@ -1,10 +1,0 @@
-package com.TLU.studentmanagement.controller;
-
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-public class LoginController {
-
-    public void setVisible(boolean b) {
-    }
-}
