--- src/com/TLU/studentmanagement/controller/RegisterController.java
@@ -1,4 +1,0 @@
-package com.TLU.studentmanagement.controller;
-
-public class RegisterController {
-}
