--- src/com/TLU/studentmanagement/util/ApiClient.java
@@ -1,6 +1,0 @@
-package com.TLU.studentmanagement.util;
-
-import org.json.JSONObject;
-
-public class ApiClient {
-}
