--- src/com/TLU/studentmanagement/view/LoginView.java
@@ -1,79 +1,0 @@
-package com.TLU.studentmanagement.view;
-
-import javax.swing.*;
-import java.awt.*;
-
-public class LoginView extends JFrame {
-    private JPanel panel;
-    private JTextField usernameField;
-    private JPasswordField passwordField;
-    private JButton loginButton;
-    private JButton registerButton;
-
-    public LoginView() {
-        setTitle("Đăng Nhập");
-        setSize(400, 300);
-        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setLocationRelativeTo(null);
-
-        // Gọi phương thức để tạo các thành phần giao diện người dùng
-        createUIComponents();
-
-        // Xử lý sự kiện nút Đăng Nhập
-
-
-        // Thêm panel vào cửa sổ chính
-        add(panel);
-    }
-
-    private void createUIComponents() {
-        // Tạo các thành phần giao diện người dùng
-        usernameField = new JTextField(20);
-        passwordField = new JPasswordField(20);
-        loginButton = new JButton("Đăng Nhập");
-        registerButton = new JButton("Đăng Ký");
-
-        // Tạo bảng điều khiển và thêm các thành phần vào bảng
-        panel = new JPanel();
-        panel.setLayout(new GridBagLayout());
-        GridBagConstraints gbc = new GridBagConstraints();
-        gbc.fill = GridBagConstraints.HORIZONTAL;
-        gbc.insets = new Insets(10, 10, 10, 10);
-
-        gbc.gridx = 0;
-        gbc.gridy = 0;
-        panel.add(new JLabel("Tên Đăng Nhập:"), gbc);
-
-        gbc.gridx = 1;
-        gbc.gridy = 0;
-        panel.add(usernameField, gbc);
-
-        gbc.gridx = 0;
-        gbc.gridy = 1;
-        panel.add(new JLabel("Mật Khẩu:"), gbc);
-
-        gbc.gridx = 1;
-        gbc.gridy = 1;
-        panel.add(passwordField, gbc);
-
-        gbc.gridx = 0;
-        gbc.gridy = 2;
-        gbc.gridwidth = 2;
-        gbc.fill = GridBagConstraints.NONE;
-        panel.add(loginButton, gbc);
-
-        gbc.gridx = 0;
-        gbc.gridy = 3;
-        gbc.gridwidth = 2;
-        gbc.fill = GridBagConstraints.NONE;
-        panel.add(registerButton, gbc);
-    }
-
-
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(() -> {
-            LoginView loginView = new LoginView();
-            loginView.setVisible(true);
-        });
-    }
-}
