+++ src/main/java/com/TLU/studentmanagement/controller/LoginController.java
@@ -1,0 +1,10 @@
+package main.java.com.TLU.studentmanagement.controller;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+public class LoginController {
+
+    public void setVisible(boolean b) {
+    }
+}
