+++ src/main/java/com/TLU/studentmanagement/controller/RegisterController.java
@@ -1,0 +1,4 @@
+package main.java.com.TLU.studentmanagement.controller;
+
+public class RegisterController {
+}
