+++ src/main/java/com/TLU/studentmanagement/main/Application.java
@@ -1,0 +1,37 @@
+package main.java.com.TLU.studentmanagement.main;
+
+
+import com.formdev.flatlaf.FlatLaf;
+import com.formdev.flatlaf.fonts.roboto.FlatRobotoFont;
+import com.formdev.flatlaf.themes.FlatMacDarkLaf;
+import main.java.com.TLU.studentmanagement.view.login.Login;
+import main.java.com.TLU.studentmanagement.manager.FormsManager;
+//import raven.toast.Notifications;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class Application extends JFrame {
+
+    public Application() {
+        init();
+    }
+
+    private void init() {
+        setTitle("FlatLaf Login");
+        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        setSize(new Dimension(1200, 700));
+        setLocationRelativeTo(null);
+        setContentPane(new Login());
+//        Notifications.getInstance().setJFrame(this);
+        FormsManager.getInstance().initApplication(this);
+    }
+
+    public static void main(String[] args) {
+        FlatRobotoFont.install();
+        FlatLaf.registerCustomDefaultsSource("resources.t3g.themes");
+        UIManager.put("defaultFont", new Font(FlatRobotoFont.FAMILY, Font.PLAIN, 13));
+        FlatMacDarkLaf.setup();
+        EventQueue.invokeLater(() -> new Application().setVisible(true));
+    }
+}
