+++ src/main/java/com/TLU/studentmanagement/manager/FormsManager.java
@@ -1,0 +1,39 @@
+package main.java.com.TLU.studentmanagement.manager;
+
+import com.formdev.flatlaf.extras.FlatAnimatedLafChange;
+import main.java.com.TLU.studentmanagement.main.Application;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class FormsManager {
+
+    private Application application;
+    private static FormsManager instance;
+
+    public static FormsManager getInstance() {
+        if (instance == null) {
+            instance = new FormsManager();
+        }
+        return instance;
+    }
+
+    private FormsManager() {
+
+    }
+
+    public void initApplication(Application application) {
+        this.application = application;
+    }
+
+    public void showForm(JComponent form) {
+        EventQueue.invokeLater(() -> {
+            FlatAnimatedLafChange.showSnapshot();
+            application.setContentPane(form);
+            application.revalidate();
+            application.repaint();
+            FlatAnimatedLafChange.hideSnapshotWithAnimation();
+        });
+    }
+
+}
