+++ src/main/java/com/TLU/studentmanagement/util/PasswordStrengthStatus.java
@@ -1,0 +1,17 @@
+package main.java.com.TLU.studentmanagement.util;
+
+import com.formdev.flatlaf.FlatClientProperties;
+import com.formdev.flatlaf.FlatLaf;
+import com.formdev.flatlaf.ui.FlatUIUtils;
+import com.formdev.flatlaf.util.UIScale;
+import net.miginfocom.swing.MigLayout;
+
+
+import javax.swing.*;
+import javax.swing.event.DocumentEvent;
+import javax.swing.event.DocumentListener;
+import java.awt.*;
+
+
+public class PasswordStrengthStatus {
+}
