+++ src/main/resources/t3g/themes/FlatDarkLaf.properties
@@ -1,0 +1,3 @@
+TitlePane.background=lighten(@background,5%)
+
+Toast.background=lighten(@background,5%)
