+++ src/main/resources/t3g/themes/FlatLaf.properties
@@ -1,0 +1,13 @@
+@accentColor=#5856D6
+
+TitlePane.unifiedBackground=false
+TitlePane.buttonSize=35,23
+
+Component.focusWidth=1
+Component.arc=10
+Button.arc=10
+Button.margin=4,6,4,6
+TextComponent.arc=10
+TextField.margin=4,6,4,6
+PasswordField.margin=4,6,4,6
+ComboBox.padding=4,6,4,6
