+++ src/main/resources/t3g/themes/FlatLightLaf.properties
@@ -1,0 +1,1 @@
+TitlePane.background=darken(@background,5%)
